# portal/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Karting Dashboard API",
        "description": "API for managing karting sessions, laps, drivers, and tracks",
        "contact": {
            "email": "support@karting-dashboard.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/tracks": {
            "get": {
                "tags": [
                    "Tracks"
                ],
                "summary": "List all tracks",
                "description": "Retrieve all tracks with their session counts",
                "operationId": "1a5aeaa66a942ddf11670313c5e107d2",
                "responses": {
                    "200": {
                        "description": "List of tracks"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Create a new track",
                "description": "Create a new karting track",
                "operationId": "a5379c7108d5509a31c095a490fec9df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "city",
                                    "country",
                                    "distance"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "city": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "country": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "distance": {
                                        "type": "number"
                                    },
                                    "corners": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Track created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/tracks/{id}": {
            "get": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Get track details",
                "description": "Retrieve a track with its sessions and laps",
                "operationId": "6e64d972925d5378792a7dc01cb1b78a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Track details"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Track not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Update a track",
                "description": "Update an existing track record",
                "operationId": "67cbfff9e245d6ba4d1064cc48773063",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Track updated"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Track not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tracks"
                ],
                "summary": "Delete a track",
                "description": "Remove a track from the system",
                "operationId": "9a46e485a6c166337897dd9e8c0734c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Track deleted"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Track not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/stats/tracks": {
            "get": {
                "tags": [
                    "Stats"
                ],
                "summary": "Get track statistics",
                "description": "Retrieve aggregated statistics for all tracks",
                "operationId": "eb69ccd7bc09a811eb6d5045f3f57e95",
                "responses": {
                    "200": {
                        "description": "Track statistics"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}